rag-agent/pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "RAG agent for RoboLearn semantic search"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "qdrant-client>=1.12.0",
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-frontmatter>=1.1.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "httpx>=0.25.0",
    "requests>=2.32.5",
    "openai-chatkit>=1.4.0",
    "openai-agents>=0.4.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "asyncpg>=0.31.0",
    "boto3>=1.41.5",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["models", "ingestion", "retrieval", "scripts"]
